Find the zero-step mismatch: two computations that were never comparable
h must be exactly 0 at theta = theta_t and was measured at RMS 0.30-0.61. The
cause is that log pi(a) came from the online forward and log pi_t(a) from the
precompute cache, whose batches were grouped differently -- and in bf16 a
response's log-probability depends on who it is batched with.

Two terms, measured apart. The accumulated sum was stored in bf16, whose spacing
is 1-2 nats at |logp| ~ 250: long responses came back as exact integers and the
batch-to-batch differences were exactly one ulp (2.0000 at 270, 1.0000 at 248,
0.0625 at 11.5), with identical token counts throughout, so never a masking
issue. Computing log_softmax with dtype=float32 fixes the accumulation and drops
the max error from 2.0000 to 1.1758. What remains is the bf16 forward's own
dependence on batch shape, RMS 0.47 -- the same order as swapping the entire
weight precision -- and no reduction precision fixes that.

So the fix is to stop comparing across batchings: a frozen copy of pi_t,
eval mode, requires_grad off, forwarded through the SAME collated batch. Never
the learner detached, never re-synced. At pi = pi_t both forwards are then
bitwise identical and h is exactly 0.

Measured: zero-step at lr 0 now gives h_abs = h_rms = h_max = 0.000000 at every
step, and the new nbpo/ref_online_minus_cache_rms diagnostic reproduces the OLD
path's h to six decimals. The whole zero-step error was this and nothing else.

A left-padding/position_ids hypothesis was tested first and rejected: supplying
position_ids changed the batch sensitivity by a factor of exactly 1.0000.

+++ scripts/simpo_trainer.py
@@ -648,7 +648,24 @@ class SimPOTrainer(Trainer):
         # dummy token; we'll ignore the losses on these tokens later
         labels[labels == label_pad_token_id] = 0
 
-        per_token_logps = torch.gather(logits.log_softmax(-1), dim=2, index=labels.unsqueeze(2)).squeeze(2)
+        # The sequence log-probability is a SUM over hundreds of response tokens and
+        # reaches magnitudes of 250-300 nats. bfloat16 has an 8-bit mantissa, so its
+        # spacing at |x| in [256, 512) is 2 and in [128, 256) is 1: accumulating this
+        # sum in bf16 quantizes the answer to +/- 1-2 nats. Measured directly -- the
+        # same response scored in two different batches differed by exactly one ulp
+        # of the stored value (2.0000 at |logp| = 270, 1.0000 at 248, 0.0625 at 11.5),
+        # which is what made h nonzero at a zero learning rate.
+        #
+        # log_softmax with an explicit dtype computes and returns float32 WITHOUT
+        # materializing a float32 copy of the (batch, seq, vocab) logits, so the
+        # gather and the reduction both run in float32 at no memory cost. This does
+        # not remove the bf16 error in the forward that produced the logits; it
+        # removes the quantization of the accumulation, which is the term that
+        # dominated the measurement.
+        per_token_logps = torch.gather(
+            logits.log_softmax(-1, dtype=torch.float32), dim=2,
+            index=labels.unsqueeze(2)).squeeze(2)
+        loss_mask = loss_mask.to(per_token_logps.dtype)
 
         if average_log_prob:
             return (per_token_logps * loss_mask).sum(-1) / loss_mask.sum(-1)
